Move VaxTestingUtil
To testutil.
